requirements.txt
# tensorflow
# matplotlib
# ipython
# numpy
# keras
# torch
# pandas
# scikit-learn
# opencv-python
# seaborn

torch==2.0.1
torchvision==0.15.2
numpy==1.26.4           
Pillow==10.0.0
matplotlib==3.7.2
scikit-learn==1.3.0
opencv-python==4.8.0.76
albumentations==1.3.1
tqdm==4.65.0   
tensorflow==2.19.0         
seaborn==0.12.2
pandas==2.0.3